requirements.txt
streamlit==1.28.0
tensorflow==2.14.0
pillow==10.0.0
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
